Organize files by what card they are on
