+++ lib/music_player/current_song_card/current_song_label.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:musbx/music_player/pick_song_button.dart';
+import 'package:musbx/music_player/current_song_card/pick_song_button.dart';
 import 'package:musbx/music_player/music_player.dart';
-import 'package:musbx/music_player/youtube_button.dart';
+import 'package:musbx/music_player/current_song_card/youtube_button.dart';
 
 class CurrentSongPanel extends StatelessWidget {
   /// Panel displaying the currently loaded song, with buttons to load a new song,
