+++ lib/music_player/loop_card/loop_slider.dart
