+++ lib/music_player/music_player_screen.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:musbx/editable_screen/card_list.dart';
-import 'package:musbx/music_player/button_panel.dart';
-import 'package:musbx/music_player/current_song_panel.dart';
+import 'package:musbx/music_player/position_card/button_panel.dart';
+import 'package:musbx/music_player/current_song_card/current_song_label.dart';
 import 'package:musbx/music_player/labeled_slider.dart';
-import 'package:musbx/music_player/loop_buttons.dart';
-import 'package:musbx/music_player/loop_slider.dart';
+import 'package:musbx/music_player/loop_card/loop_buttons.dart';
+import 'package:musbx/music_player/loop_card/loop_slider.dart';
 import 'package:musbx/music_player/music_player.dart';
-import 'package:musbx/music_player/position_slider.dart';
+import 'package:musbx/music_player/position_card/position_slider.dart';
 
 class MusicPlayerScreen extends StatelessWidget {
   /// Screen that allows the user to select and play a song.
